requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==2.2.4
django-cors-headers==3.0.2
django-dynamic-fixture==2.0.0
django-extensions==2.1.9
django-filter==2.1.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.9.4
djoser==1.7.0
ipdb==0.12.2
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Markdown==3.1.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
